wordSquad/requirements.txt
Django==4.0.6
psycopg2-binary==2.9.3
python-telegram-bot==13.13
nltk==3.7
